+++ crates/cxx-qt-gen/test_outputs/signals.rs
@@ -208,14 +208,4 @@ pub mod cxx_qt_ffi {
             self.cxx_qt_ffi_rust_mut()
         }
     }
-    #[doc = r" Generated CXX-Qt module containing type alias to the C++ types of the QObjects"]
-    pub mod qobject {
-        #[doc = "The C++ type for the QObject "]
-        #[doc = "MyObject"]
-        #[doc = "\n"]
-        #[doc = "Use this type when referring to the QObject as a pointer"]
-        #[doc = "\n"]
-        #[doc = "See the book for more information: <https://kdab.github.io/cxx-qt/book/qobject/generated-qobject.html>"]
-        pub type MyObject = super::MyObject;
-    }
 }